Small refactorings
// Support block aliases
// Bug fix - Json String as list, fix parsing numbers instead lazily parsed numbers
// 0 REFACTOR Port - evaluate if process() is not called too often, e.g. double call onIncomingConnectionAdded and incomingDataChanged, also on onConnectionRemoved
//      - Trigger only when connections are added or removed and data changed. In case data changed and a connection was removed or added, process safely is only called once.

+++ src/main/java/vpllib/method/JsonMethods.java
@@ -50,7 +50,7 @@ public class JsonMethods {
                     if (primitive.isString()) {
                         list.add(primitive.getAsString());
                     } else if (primitive.isNumber()) {
-                        Number number = primitive.getAsNumber();
+                        Number number = primitive.getAsNumber().doubleValue();
                         list.add(ParsingUtils.castToBestNumericType(number));
                     } else if (primitive.isBoolean()) {
                         list.add(primitive.getAsBoolean());
